refactor(runtime): extract inline runtime tests to dedicated module (#3192)
